refactor(web): share one provider config form across onboarding + settings modal
Extract the credential form into a single `ProviderConfigForm` reused by both the
onboarding wizard and the settings "Add provider" modal, so they can't drift.
Each surface owns only its own chrome:

- Onboarding (`SetupWizard`): keeps the welcome heading, 2-step indicator and
  "Set up an AI provider" blurb wrapping the shared picker.
- Settings modal (`AddAiProviderDialog`): just its title, back affordance and
  close button — no step indicator, no onboarding blurb.

The shared form renders only the inputs the selected provider supports (API key
or subscription paste) plus an optional, settings-only Name field; the submit
label is caller-supplied ("Add provider" vs "Save").
